refactor: use path aliases for cleaner imports
