+++ src/components/ui/badge.jsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { cva } from "class-variance-authority";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const badgeVariants = cva(
   "inline-flex items-center rounded-md border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
