+++ src/components/ui/card.jsx
@@ -1,6 +1,5 @@
 import * as React from "react";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Card = React.forwardRef(({ className, ...props }, ref) => (
   <div
