+++ src/components/ui/checkbox.jsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from "react";
 import { motion } from "framer-motion";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Checkbox = forwardRef(
   ({ className, checked, onCheckedChange, disabled, ...props }, ref) => {
