+++ src/components/ui/progress.jsx
@@ -1,9 +1,7 @@
 "use client";
-
 import * as React from "react";
 import * as ProgressPrimitive from "@radix-ui/react-progress";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Progress = React.forwardRef(({ className, value, ...props }, ref) => (
   <ProgressPrimitive.Root
