+++ src/components/ui/radio-group.jsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import * as RadioGroupPrimitive from "@radix-ui/react-radio-group";
 import { Circle } from "lucide-react";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const RadioGroup = React.forwardRef(({ className, ...props }, ref) => {
   return (
