+++ src/components/ui/select.jsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown, ChevronUp } from "lucide-react";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Select = SelectPrimitive.Root;
