+++ src/components/ui/separator.jsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import * as SeparatorPrimitive from "@radix-ui/react-separator";
-
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Separator = React.forwardRef(
   ({ className, orientation = "horizontal", decorative = true, ...props }, ref) => (
