+++ src/components/ui/skeleton.jsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 function Skeleton({ className, ...props }) {
   return (
