+++ src/components/ui/switch.jsx
@@ -1,8 +1,7 @@
 "use client";
-
 import * as React from "react";
 import * as SwitchPrimitives from "@radix-ui/react-switch";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Switch = React.forwardRef(({ className, ...props }, ref) => (
   <SwitchPrimitives.Root
