+++ src/components/ui/tabs.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as TabsPrimitive from "@radix-ui/react-tabs";
-import { cn } from "../../lib/utils";
+import { cn } from "@/lib/utils";
 
 const Tabs = TabsPrimitive.Root;
