+++ src/main.jsx
@@ -1,10 +1,10 @@
+import { initializeConsole } from "@/lib/consoleIntro.js";
+import { analytics } from "@/services/analyticsService";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./app";
 import "./app.css";
 import "./i18n";
-import { analytics } from "./services/analyticsService";
-import { initializeConsole } from "./lib/consoleIntro.js";
 
 window.addEventListener("beforeunload", () => {
   analytics.flushEvents();
